backend/requirements.txt
fastapi
uvicorn
opencv-python
pyzbar
Pillow
pandas
openpyxl
python-multipart
google-genai